requirements.txt
beautifulsoup4==4.11.2
certifi==2022.9.24
charset-normalizer==2.1.1
config==0.5.1
contourpy==1.0.6
cycler==0.11.0
et-xmlfile==1.1.0
fonttools==4.38.0
google==3.0.0
google-cloud==0.34.0
idna==3.4
kiwisolver==1.4.4
matplotlib==3.6.2
numpy==1.23.5
openpyxl==3.0.10
packaging==21.3
paho-mqtt==1.6.1
pandas==1.5.2
Pillow==9.3.0
protobuf==4.21.12
pyorbital==1.7.3
pyparsing==3.0.9
pyserial==3.5
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
requests==2.28.1
scipy==1.10.1
six==1.16.0
soupsieve==2.3.2.post1
urllib3==1.26.13
skyfield==1.45